✨ refactor ProjectView to separate module
Moved ProjectView and its components to a new file for modularity.
